# backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Board Game Prototype API",
    "version": "1.0.0"
  },
  "paths": {
    "/auth/google": {
      "get": {
        "summary": "Googleログイン",
        "description": "Googleアカウントを使用してログインします。",
        "responses": {
          "302": {
            "description": "リダイレクト",
            "headers": {
              "Location": {
                "description": "リダイレクト先のURL",
                "schema": {
                  "type": "string"
                },
                "example": "https://accounts.google.com/o/oauth2/auth"
              }
            }
          }
        }
      }
    },
    "/auth/google/callback": {
      "get": {
        "summary": "Googleログインコールバック",
        "description": "GoogleログインのコールバックURL。",
        "responses": {
          "302": {
            "description": "ログイン成功時にリダイレクト",
            "headers": {
              "Location": {
                "description": "リダイレクト先のURL",
                "schema": {
                  "type": "string"
                },
                "example": "http://localhost:3000/prototypes"
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "ログアウト",
        "description": "現在のセッションを終了し、ユーザーをログアウトします。",
        "responses": {
          "200": {
            "description": "ログアウト成功",
            "content": {
              "application/json": {
                "example": {
                  "message": "Logged out successfully"
                }
              }
            }
          },
          "500": {
            "description": "ログアウト失敗",
            "content": {
              "application/json": {
                "example": {
                  "error": "Failed to logout"
                }
              }
            }
          }
        }
      }
    },
    "/api/prototypes": {
      "get": {
        "summary": "プロトタイプ一覧取得",
        "description": "ユーザーがアクセス可能なプロトタイプの一覧を取得します。",
        "responses": {
          "200": {
            "description": "アクセス可能なプロトタイプの一覧を返します",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "PrototypeModel"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "プロトタイプ作成",
        "description": "新しいプロトタイプを作成します。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "playerCount": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "新しいプロトタイプを作成しました",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": "PrototypeModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "リクエストが不正です"
          },
          "500": {
            "description": "サーバーエラー"
          }
        }
      }
    },
    "/api/prototypes/{prototypeId}": {
      "get": {
        "summary": "プロトタイプ取得",
        "description": "指定されたIDのプロトタイプを取得します。",
        "parameters": [
          {
            "name": "prototypeId",
            "in": "path",
            "required": true,
            "description": "プロトタイプのID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "プロトタイプの詳細を返します",
            "content": {
              "application/json": {
                "schema": {
                  "type": "PrototypeModel"
                }
              }
            }
          },
          "404": {
            "description": "プロトタイプが見つかりません"
          }
        }
      },
      "put": {
        "summary": "プロトタイプ更新",
        "description": "指定されたIDのプロトタイプを更新します。",
        "parameters": [
          {
            "name": "prototypeId",
            "in": "path",
            "required": true,
            "description": "プロトタイプのID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "プロトタイプを更新しました",
            "content": {
              "application/json": {
                "schema": {
                  "type": "PrototypeModel"
                }
              }
            }
          },
          "404": {
            "description": "プロトタイプが見つかりません"
          }
        }
      },
      "delete": {
        "summary": "プロトタイプ削除",
        "description": "指定されたIDのプロトタイプを削除します。",
        "parameters": [
          {
            "name": "prototypeId",
            "in": "path",
            "required": true,
            "description": "プロトタイプのID",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "プロトタイプを削除しました",
            "content": {
              "application/json": {
                "schema": {
                  "type": "PrototypeModel"
                }
              }
            }
          },
          "404": {
            "description": "プロトタイプが見つかりません"
          }
        }
      }
    },
    "/api/prototypes/{prototypeId}/versions": {
      "get": {
        "summary": "プロトタイプバージョン一覧取得",
        "description": "指定されたプロトタイプのバージョン一覧を取得します。",
        "parameters": [
          {
            "name": "prototypeId",
            "in": "path",
            "required": true,
            "description": "プロトタイプのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "プロトタイプのバージョン一覧を返します",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "prototype": {
                      "type": "PrototypeModel"
                    },
                    "versions": {
                      "type": "array",
                      "items": {
                        "type": "PrototypeVersionModel"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "プロトタイプが見つかりません"
          }
        }
      }
    },
    "/api/prototypes/groups/{groupId}": {
      "get": {
        "summary": "グループのプロトタイプ一覧取得",
        "description": "指定されたグループに属するプロトタイプの一覧を取得します。",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "description": "グループのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "グループに属するプロトタイプの一覧を返します",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "PrototypeModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/prototypes/groups/{groupId}/accessUsers": {
      "get": {
        "summary": "グループへのアクセス権を取得",
        "description": "指定されたグループにアクセス可能なユーザーを取得します。",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "description": "グループのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "アクセス可能なユーザーの一覧を返します",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "UserModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/prototypes/groups/{groupId}/invite": {
      "post": {
        "summary": "ユーザーにプロトタイプへのアクセス権を付与",
        "description": "指定されたグループにユーザーを招待します。",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "description": "グループのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "guestIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ユーザーを招待しました"
          },
          "400": {
            "description": "リクエストが不正です"
          },
          "404": {
            "description": "グループが見つかりません"
          },
          "500": {
            "description": "サーバーエラー"
          }
        }
      }
    },
    "/api/prototypes/groups/{groupId}/invite/{guestId}": {
      "delete": {
        "summary": "ユーザーのアクセス権を削除",
        "description": "指定されたグループからユーザーのアクセス権を削除します。",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "description": "グループのID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "guestId",
            "in": "path",
            "required": true,
            "description": "ゲストユーザーのID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ユーザーのアクセス権を削除しました"
          },
          "400": {
            "description": "リクエストが不正です"
          },
          "404": {
            "description": "グループが見つかりません"
          },
          "500": {
            "description": "サーバーエラー"
          }
        }
      }
    },
    "/user": {
      "get": {
        "summary": "ユーザー情報取得",
        "description": "現在ログインしているユーザーの情報を取得します。",
        "responses": {
          "200": {
            "description": "ユーザー情報を返します",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    }
                  }
                },
                "examples": {
                  "authenticated": {
                    "summary": "認証済みユーザー",
                    "value": {
                      "id": 1,
                      "username": "exampleUser"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/search": {
      "get": {
        "summary": "ユーザー検索",
        "description": "ユーザー名でユーザーを検索します。",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "required": true,
            "description": "検索するユーザー名",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "検索結果を返します",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "username": {
                        "type": "string"
                      }
                    }
                  }
                },
                "examples": {
                  "search": {
                    "summary": "検索結果",
                    "value": [
                      {
                        "id": 1,
                        "googleId": "exampleGoogleId",
                        "username": "exampleUser"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}